fix issue #3: don't force clean builds
Clean builds can be manually enforced via

  make CLEAN=yes ENGINE=<engine>

+++ build-mms.sh
@@ -1,18 +1,29 @@
 #!/bin/bash
 
+set -e
+
 sdk=${1}
 current_git_branch=$(git branch --show-current)
 patches_dir=patches/linux/gcc
 
+clean_build=${2}
 
-rm -rf hl2sdk-${sdk} metamod-source
-git submodule update --init --recursive
+if [[ ${clean_build} ]]; then
+  rm -rf output/*
+  rm -rf metamod-source/build
+  rm -rf hl2sdk-${sdk} metamod-source
+
+  git submodule update --init --recursive
+fi
 
 cd hl2sdk-${sdk}
+rm -rf ../.git/modules/$(basename $(readlink -f .))/rebase-apply
+
+git_ref=$(git rev-parse HEAD)
+git reset --hard
+git checkout ${git_ref}
 git branch -D temp
 git checkout -b temp
-
-rm -rf ../.git/modules/$(basename $(readlink -f .))/rebase-apply
 
 # apply patches
 for git_patch in $(ls ../${patches_dir}/hl2sdk-${sdk}/*.patch); do
@@ -24,20 +35,20 @@ git checkout ${current_git_branch}
 
 mkdir -p metamod-source/build
 cd metamod-source/build
+rm -rf ../.git/modules/$(basename $(readlink -f .))/rebase-apply
+
+git_ref=$(git rev-parse HEAD)
+git reset --hard
+git checkout ${git_ref}
+git branch -D temp
+git checkout -b temp
 
 python3 -m venv .venv
 source .venv/bin/activate
 pip install wheel
 
-rm -rf /tmp/.srcds-tickrate-enabler
-git clone https://github.com/alliedmodders/ambuild /tmp/.srcds-tickrate-enabler/ambuild
-pip install /tmp/.srcds-tickrate-enabler/ambuild
-rm -rf /tmp/.srcds-tickrate-enabler/ambuild
-
-rm -rf ../.git/modules/$(basename $(readlink -f .))/rebase-apply
-git submodule update --init
-git branch -D temp
-git checkout -b temp
+test -d .ambuild-git || git clone https://github.com/alliedmodders/ambuild .ambuild-git
+pip install .ambuild-git
 
 # apply patches
 for git_patch in $(ls ../../${patches_dir}/metamod-source/*.patch); do
